+++ src/javascripts/components/views/eventDetailsView.js
@@ -107,10 +107,10 @@ const eventDetailsView = (eventFirebaseKey) => {
         });
       });
       Promise.all([
-        eventData.foodTotalPrices(eventFirebaseKey),
-        eventData.showsTotalPrices(eventFirebaseKey),
-        eventData.souvenirsTotalPrices(eventFirebaseKey),
-        eventData.staffTotalPrices(eventFirebaseKey),
+        eventFood.foodTotalPrices(eventFirebaseKey),
+        eventShows.showsTotalPrices(eventFirebaseKey),
+        eventSouvenirs.souvenirsTotalPrices(eventFirebaseKey),
+        eventStaff.staffTotalPrices(eventFirebaseKey),
       ]).then((values) => {
         const eventTotal = values.reduce((a, b) => a + b);
         $('#eventTotal').html(
@@ -126,68 +126,38 @@ const eventDetailsView = (eventFirebaseKey) => {
     const firebaseKey = e.currentTarget.id;
     $(`.line-item#${firebaseKey}`).remove();
     eventFood.deleteFoodOfEvent(firebaseKey);
-    Promise.all([
-      eventData.foodTotalPrices(eventFirebaseKey),
-      eventData.showsTotalPrices(eventFirebaseKey),
-      eventData.souvenirsTotalPrices(eventFirebaseKey),
-      eventData.staffTotalPrices(eventFirebaseKey),
-    ]).then((values) => {
-      const eventTotal = values.reduce((a, b) => a + b);
-      $('#eventTotal').html(
-        `<h2 id="eventTotalBanner"> The Total Cost is ${eventTotal}</h2>`
-      );
-    });
+    // Promise.all(eventFood.foodTotalPrices(eventFirebaseKey)).then((foodPrice) => {
+    //   $('#foodTotalCost').html(`${foodPrice}`);
+    // });
+    // Promise.all([
+    //   eventFood.foodTotalPrices(eventFirebaseKey),
+    //   eventShows.showsTotalPrices(eventFirebaseKey),
+    //   eventSouvenirs.souvenirsTotalPrices(eventFirebaseKey),
+    //   eventStaff.staffTotalPrices(eventFirebaseKey),
+    // ]).then((values) => {
+    //   const eventTotal = values.reduce((a, b) => a + b);
+    //   $('#eventTotal').html(
+    //     `<h2 id="eventTotalBanner"> The Total Cost is ${eventTotal}</h2>`
+    //   );
+    // });
   });
   $('body').on('click', '.delete-event-show', (e) => {
     e.stopImmediatePropagation();
     const firebaseKey = e.currentTarget.id;
     $(`.line-item#${firebaseKey}`).remove();
     eventShows.deleteShowsOfEvent(firebaseKey);
-    Promise.all([
-      eventData.foodTotalPrices(eventFirebaseKey),
-      eventData.showsTotalPrices(eventFirebaseKey),
-      eventData.souvenirsTotalPrices(eventFirebaseKey),
-      eventData.staffTotalPrices(eventFirebaseKey),
-    ]).then((values) => {
-      const eventTotal = values.reduce((a, b) => a + b);
-      $('#eventTotal').html(
-        `<h2 id="eventTotalBanner"> The Total Cost is ${eventTotal}</h2>`
-      );
-    });
   });
   $('body').on('click', '.delete-event-souvenir', (e) => {
     e.stopImmediatePropagation();
     const firebaseKey = e.currentTarget.id;
     $(`.line-item#${firebaseKey}`).remove();
     eventSouvenirs.deleteSouvenirsOfEvent(firebaseKey);
-    Promise.all([
-      eventData.foodTotalPrices(eventFirebaseKey),
-      eventData.showsTotalPrices(eventFirebaseKey),
-      eventData.souvenirsTotalPrices(eventFirebaseKey),
-      eventData.staffTotalPrices(eventFirebaseKey),
-    ]).then((values) => {
-      const eventTotal = values.reduce((a, b) => a + b);
-      $('#eventTotal').html(
-        `<h2 id="eventTotalBanner"> The Total Cost is ${eventTotal}</h2>`
-      );
-    });
   });
   $('body').on('click', '.delete-event-staff', (e) => {
     e.stopImmediatePropagation();
     const firebaseKey = e.currentTarget.id;
     $(`.line-item#${firebaseKey}`).remove();
     eventStaff.deleteStaffOfEvent(firebaseKey);
-    Promise.all([
-      eventData.foodTotalPrices(eventFirebaseKey),
-      eventData.showsTotalPrices(eventFirebaseKey),
-      eventData.souvenirsTotalPrices(eventFirebaseKey),
-      eventData.staffTotalPrices(eventFirebaseKey),
-    ]).then((values) => {
-      const eventTotal = values.reduce((a, b) => a + b);
-      $('#eventTotal').html(
-        `<h2 id="eventTotalBanner"> The Total Cost is ${eventTotal}</h2>`
-      );
-    });
   });
 };
