+++ src/javascripts/helpers/data/event_souvenirs.js
@@ -1,5 +1,6 @@
 import axios from 'axios';
 import apiKeys from '../apiKeys.json';
+import souvenirsData from './souvenirsData';
 
 const baseUrl = apiKeys.firebaseKeys.databaseURL;
 
@@ -29,4 +30,23 @@ const addSouvenirsOfEvents = (dataObject) => {
   }).catch((error) => console.warn(error));
 };
 
-export default { getEventSouvenirs, addSouvenirsOfEvents, deleteSouvenirsOfEvent };
+const souvenirsTotalPrices = (eventFirebaseKey) => new Promise((resolve, reject) => {
+  let souvenirsTotal = 0;
+  getEventSouvenirs(eventFirebaseKey)
+    // execute getSinglesouvenirs for all elements in the array and resolve when all are resolved
+    .then((souvenirsArray) => Promise.all(souvenirsArray.map((souvenirs) => souvenirsData.getSingleSouvenir(souvenirs.souvenirUid))))
+    // Add up the the prices to variable
+    .then((souvenirsObjects) => souvenirsObjects.forEach((souvenir) => {
+      souvenirsTotal += parseInt(souvenir.price, 10);
+    }))
+    // resolve the promise with the final total
+    .then(() => resolve(souvenirsTotal))
+    .catch((error) => reject(error));
+});
+
+export default {
+  getEventSouvenirs,
+  addSouvenirsOfEvents,
+  deleteSouvenirsOfEvent,
+  souvenirsTotalPrices
+};
